chore: seperate Tag dataclasses from tags cog for imports
